Update API lambda log group name to be more readable
Prior log group name was a long truncated auto-generated
name that had a long set of random alphanumeric characters,
which would make it difficult to identify and search for
if there were many other Lambda functions with CloudWatch
Logs.

Specified a log group with the more specific name
"ConsentManagementApi-Lambda-ApplicationLogs", so we can
instantly identify which CloudWatch log group hosts the
application logs for the Consent Management API service's
Lambda function.