Remove Generate API Clients workflow
Found that the Java API client auto-generated
by AWS API Gateway is not functional out of the
box and is using very old deprecated libraries
that will be a time sink to upgrade, so will likely
but more efficient to start from OpenAPI Generator
generated code.

Dropping the Generate API Clients GitHub Workflow
and will mark the resulting GitHub repository as
work in progress for now, not critical to have
at this time.