Add prebuild step to sync latest Open API spec
Update the CDK package to have a prebuild script that
syncs the local resources/ version of the Consent
Management API OpenAPI JSON spec with the latest
from the API models package.

Found that this is significantly simpler than trying
to update the CDK TypeScript code itself to fetch
the file entirely dynamically, since fetching data
is generally an async operation that propagates async
methods up to the caller, while CDK stack constructors
must be synchronous.

This update also results in automatically failing the
"Build" GitHub Workflow when there are API model
changes until the code repo is updated to have the
latest spec file, since test CDK stack snapshots will
be out of sync if the file doesn't match the latest
from the models file.  Intentionally keeping this
behaviour so we block CD/CI deployments when our
CDK package isn't in sync with the latest API models.