Add consentVersion to ActiveConsentsWithExpiryTime GSI
In order to expire consents via the Consent Management API,
callers must provide consentVersion, with the API service
making a conditional update request that fails if the
existing consent version is not equal to the provided
consent version minus one.

This guards against concurrent writes that could otherwise
result in unexpected/corrupted data.

Since we need this attribute to update consents, adding
it to the GSI used by the Consent Expiry Processor to
identify consents that should be expired.