Update ActiveConsentsWithExpiryTime GSI partition key name
Update the partition key attribute name for the
ActiveConsentsWithExpiryTime GSI from activeId to
autoExpireId to more accurately reflect its use
case and reduce later confusion for developers.

This attribute should be populated with the consent's
partition key value if and only if both of the following
are true:
1. The consent currently has consentStatus = "ACTIVE"; and
2. The consent has a non-null expiryTime value.

This ensures that the auto expiry GSI only includes the
consent if it's in-scope for auto expiry, without requiring
us to wipe the expiryTime after expiring the consent, since
it's relevant to keep expiryTime for users to be able to
query afterwards.

The prior name of activeId was misleading since we do not
want to populate this attribute for all active consents,
only for active consents with expiryTime values.